Core Abstract Functions of the App

+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/CommandHandlers/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Request.Commands;
 using HR.LeaveManagement.Application.Persistance.Contracts;
 using MediatR;
@@ -24,6 +25,9 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.CommandHan
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+                throw new NotFoundException(nameof(leaveType), request.Id);
 
             await _leaveTypeRepository.Delete(leaveType);
